requirements.txt
pinecone-client[grpc]
sentence-transformers
pandas
torch
gradio
tqdm
transformers
accelerate
tqdm
openpyxl